Create Mux_n_1Bit
It implements a multiplexer that selects one bit from the Data_in input based on the value of sel and assigns it to the output f. The for loop iterates over the possible values of sel and checks if it matches the loop index i. If there is a match, the corresponding bit from Data_in is assigned to f.